# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TransBus - Sistema de rastreo de transporte público
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Route {
  id: number;
  name: string;
  description?: string | null;
  color: string;
  active: boolean;
  createdAt: string;
}

export interface CreateRoute {
  name: string;
  description?: string | null;
  color?: string;
}

export interface UpdateRoute {
  name?: string;
  description?: string | null;
  color?: string;
  active?: boolean;
}

export interface Stop {
  id: number;
  routeId: number;
  name: string;
  lat: number;
  lng: number;
  order: number;
  createdAt: string;
}

export interface CreateStop {
  name: string;
  lat: number;
  lng: number;
  order: number;
}

export interface DeletedResult {
  ok: boolean;
  id: number;
}

export interface Vehicle {
  id: number;
  routeId: number;
  driverName: string;
  plateNumber: string;
  pin: string;
  active: boolean;
  createdAt: string;
}

export interface CreateVehicle {
  routeId: number;
  driverName: string;
  plateNumber: string;
  /** PIN de 4 dígitos para autenticar al chofer */
  pin: string;
}

export interface UpdateVehicle {
  routeId?: number;
  driverName?: string;
  plateNumber?: string;
  pin?: string;
  active?: boolean;
}

export interface VerifyPin {
  pin: string;
}

export interface PinResult {
  ok: boolean;
  vehicleId: number;
}

export interface UpdatePosition {
  lat: number;
  lng: number;
  speed?: number | null;
  heading?: number | null;
}

export interface PositionAck {
  ok: boolean;
  recordedAt: string;
}

export interface StopEta {
  stopId: number;
  stopName: string;
  lat: number;
  lng: number;
  order: number;
  etaMinutes?: number | null;
  distanceKm?: number | null;
}

export interface EtaResponse {
  vehicleId: number;
  currentLat: number;
  currentLng: number;
  stops: StopEta[];
}

export interface VehicleLive {
  vehicleId: number;
  driverName: string;
  plateNumber: string;
  lat: number;
  lng: number;
  speed?: number | null;
  heading?: number | null;
  updatedAt?: string | null;
}

export interface RouteDetail {
  id: number;
  name: string;
  description?: string | null;
  color: string;
  active: boolean;
  stops: Stop[];
  vehicles: Vehicle[];
}

export interface RouteLive {
  routeId: number;
  vehicles: VehicleLive[];
}

export interface OfflineAlert {
  shiftId: number;
  shiftStartedAt: string;
  vehicleId: number;
  plateNumber: string;
  driverName: string;
  routeId: number;
  lastPositionAt?: string | null;
  minutesSilent?: number | null;
}

export interface CreateShift {
  vehicleId: number;
}

export interface ShiftRecord {
  id: number;
  vehicleId: number;
  startedAt: string;
  endedAt?: string | null;
  positionsCount: number;
}

export interface ShiftDetail {
  id: number;
  vehicleId: number;
  startedAt: string;
  endedAt?: string | null;
  positionsCount: number;
  plateNumber: string;
  driverName: string;
}

export interface PositionPoint {
  lat: number;
  lng: number;
  speed?: number | null;
  heading?: number | null;
  recordedAt: string;
}

export interface ShiftPositions {
  shiftId: number;
  vehicleId: number;
  plateNumber: string;
  driverName: string;
  startedAt: string;
  endedAt?: string | null;
  positions: PositionPoint[];
}

export type ListOfflineVehiclesParams = {
thresholdMin?: number;
};

export type ListShiftsParams = {
vehicleId?: number;
};