# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransBus - Sistema de rastreo de transporte público
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: routes
    description: Rutas de transporte
  - name: stops
    description: Paradas de una ruta
  - name: vehicles
    description: Vehículos / choferes
  - name: positions
    description: Posiciones GPS en tiempo real
  - name: alerts
    description: Alertas de vehículos sin señal

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: Lista todas las rutas
      responses:
        "200":
          description: Lista de rutas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Crea una nueva ruta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoute"
      responses:
        "201":
          description: Ruta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"

  /routes/{id}:
    get:
      operationId: getRoute
      tags: [routes]
      summary: Obtiene una ruta con sus paradas y vehículos activos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detalle de la ruta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteDetail"
        "404":
          description: No encontrada
    put:
      operationId: updateRoute
      tags: [routes]
      summary: Edita una ruta existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoute"
      responses:
        "200":
          description: Ruta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "404":
          description: No encontrada
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Elimina una ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResult"
        "404":
          description: No encontrada

  /routes/{id}/stops:
    get:
      operationId: listStops
      tags: [stops]
      summary: Lista las paradas de una ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de paradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stop"
    post:
      operationId: createStop
      tags: [stops]
      summary: Agrega una parada a la ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStop"
      responses:
        "201":
          description: Parada creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"

  /routes/{id}/stops/{stopId}:
    delete:
      operationId: deleteStop
      tags: [stops]
      summary: Elimina una parada de la ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResult"
        "404":
          description: No encontrada

  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: Lista todos los vehículos
      responses:
        "200":
          description: Lista de vehículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Registra un nuevo vehículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicle"
      responses:
        "201":
          description: Vehículo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  /vehicles/{id}:
    put:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Edita un vehículo existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicle"
      responses:
        "200":
          description: Vehículo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: No encontrado
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Elimina un vehículo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResult"
        "404":
          description: No encontrado

  /vehicles/{id}/verify-pin:
    post:
      operationId: verifyPin
      tags: [vehicles]
      summary: Verifica el PIN de un vehículo antes de iniciar turno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPin"
      responses:
        "200":
          description: PIN correcto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinResult"
        "401":
          description: PIN incorrecto

  /vehicles/{id}/position:
    post:
      operationId: updatePosition
      tags: [positions]
      summary: Actualiza la posición GPS de un vehículo (app del chofer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePosition"
      responses:
        "200":
          description: Posición registrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionAck"

  /vehicles/{id}/eta:
    get:
      operationId: getEta
      tags: [positions]
      summary: Obtiene ETA del vehículo a cada parada de su ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ETAs por parada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EtaResponse"

  /vehicles/offline:
    get:
      operationId: listOfflineVehicles
      tags: [alerts]
      summary: Vehículos con turno activo que dejaron de reportar GPS
      parameters:
        - name: thresholdMin
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de alertas de vehículos sin señal
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfflineAlert"

  /shifts:
    post:
      operationId: createShift
      tags: [shifts]
      summary: Inicia un nuevo turno
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShift"
      responses:
        "201":
          description: Turno iniciado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftRecord"
    get:
      operationId: listShifts
      tags: [shifts]
      summary: Lista turnos (opcionalmente filtrado por vehículo)
      parameters:
        - name: vehicleId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Lista de turnos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShiftDetail"

  /shifts/{id}/end:
    patch:
      operationId: endShift
      tags: [shifts]
      summary: Finaliza un turno activo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Turno finalizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftRecord"
        "404":
          description: Turno no encontrado

  /shifts/{id}/positions:
    get:
      operationId: getShiftPositions
      tags: [shifts]
      summary: Historial de posiciones GPS de un turno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de posiciones del turno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftPositions"
        "404":
          description: Turno no encontrado

  /routes/{id}/live:
    get:
      operationId: getRouteLive
      tags: [positions]
      summary: Posiciones actuales de todos los vehículos en una ruta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posiciones en tiempo real
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteLive"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, color, active, createdAt]

    CreateRoute:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
      required: [name]

    UpdateRoute:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        active:
          type: boolean

    Stop:
      type: object
      properties:
        id:
          type: integer
        routeId:
          type: integer
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, routeId, name, lat, lng, order, createdAt]

    CreateStop:
      type: object
      properties:
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        order:
          type: integer
      required: [name, lat, lng, order]

    DeletedResult:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
      required: [ok, id]

    Vehicle:
      type: object
      properties:
        id:
          type: integer
        routeId:
          type: integer
        driverName:
          type: string
        plateNumber:
          type: string
        pin:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, routeId, driverName, plateNumber, pin, active, createdAt]

    CreateVehicle:
      type: object
      properties:
        routeId:
          type: integer
        driverName:
          type: string
        plateNumber:
          type: string
        pin:
          type: string
          description: PIN de 4 dígitos para autenticar al chofer
      required: [routeId, driverName, plateNumber, pin]

    UpdateVehicle:
      type: object
      properties:
        routeId:
          type: integer
        driverName:
          type: string
        plateNumber:
          type: string
        pin:
          type: string
        active:
          type: boolean

    VerifyPin:
      type: object
      properties:
        pin:
          type: string
      required: [pin]

    PinResult:
      type: object
      properties:
        ok:
          type: boolean
        vehicleId:
          type: integer
      required: [ok, vehicleId]

    UpdatePosition:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
          nullable: true
        heading:
          type: number
          nullable: true
      required: [lat, lng]

    PositionAck:
      type: object
      properties:
        ok:
          type: boolean
        recordedAt:
          type: string
          format: date-time
      required: [ok, recordedAt]

    StopEta:
      type: object
      properties:
        stopId:
          type: integer
        stopName:
          type: string
        lat:
          type: number
        lng:
          type: number
        order:
          type: integer
        etaMinutes:
          type: number
          nullable: true
        distanceKm:
          type: number
          nullable: true
      required: [stopId, stopName, lat, lng, order]

    EtaResponse:
      type: object
      properties:
        vehicleId:
          type: integer
        currentLat:
          type: number
        currentLng:
          type: number
        stops:
          type: array
          items:
            $ref: "#/components/schemas/StopEta"
      required: [vehicleId, currentLat, currentLng, stops]

    VehicleLive:
      type: object
      properties:
        vehicleId:
          type: integer
        driverName:
          type: string
        plateNumber:
          type: string
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
          nullable: true
        heading:
          type: number
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required: [vehicleId, driverName, plateNumber, lat, lng]

    RouteDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        active:
          type: boolean
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
      required: [id, name, color, active, stops, vehicles]

    RouteLive:
      type: object
      properties:
        routeId:
          type: integer
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/VehicleLive"
      required: [routeId, vehicles]

    OfflineAlert:
      type: object
      properties:
        shiftId:
          type: integer
        shiftStartedAt:
          type: string
          format: date-time
        vehicleId:
          type: integer
        plateNumber:
          type: string
        driverName:
          type: string
        routeId:
          type: integer
        lastPositionAt:
          type: string
          format: date-time
          nullable: true
        minutesSilent:
          type: integer
          nullable: true
      required: [shiftId, shiftStartedAt, vehicleId, plateNumber, driverName, routeId]

    CreateShift:
      type: object
      properties:
        vehicleId:
          type: integer
      required: [vehicleId]

    ShiftRecord:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        positionsCount:
          type: integer
      required: [id, vehicleId, startedAt, positionsCount]

    ShiftDetail:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        positionsCount:
          type: integer
        plateNumber:
          type: string
        driverName:
          type: string
      required: [id, vehicleId, startedAt, positionsCount, plateNumber, driverName]

    PositionPoint:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
          nullable: true
        heading:
          type: number
          nullable: true
        recordedAt:
          type: string
          format: date-time
      required: [lat, lng, recordedAt]

    ShiftPositions:
      type: object
      properties:
        shiftId:
          type: integer
        vehicleId:
          type: integer
        plateNumber:
          type: string
        driverName:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        positions:
          type: array
          items:
            $ref: "#/components/schemas/PositionPoint"
      required: [shiftId, vehicleId, plateNumber, driverName, startedAt, positions]