# lib/api-zod/src/generated/types/updateRoute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TransBus - Sistema de rastreo de transporte público
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateRoute {
  name?: string;
  description?: string | null;
  color?: string;
  active?: boolean;
}