# lib/api-zod/src/generated/types/updateVehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TransBus - Sistema de rastreo de transporte público
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateVehicle {
  routeId?: number;
  driverName?: string;
  plateNumber?: string;
  pin?: string;
  active?: boolean;
}